// contexts/events/types.ts
enum Types {
  CREATE_EVENT = "CREATE_EVENT",
  CREATE_EVENT_ERROR = "CREATE_EVENT_ERROR",
  CREATE_EVENT_SUCCESS = "CREATE_EVENT_SUCCESS",

  DELETE_EVENT = "DELETE_EVENT",
  DELETE_EVENT_ERROR = "DELETE_EVENT_ERROR",
  DELETE_EVENT_SUCCESS = "DELETE_EVENT_SUCCESS",

  UPDATE_EVENT = "UPDATE_EVENT",
  UPDATE_EVENT_ERROR = "UPDATE_EVENT_ERROR",
  UPDATE_EVENT_SUCCESS = "UPDATE_EVENT_SUCCESS",

  RESET_NEW_EVENT = "RESET_NEW_EVENT",
  SET_EVENTS = "SET_EVENTS",

  GET_EVENTS = "GET_EVENTS",
  GET_EVENTS_ERROR = "GET_EVENTS_ERROR",
  GET_EVENTS_SUCCESS = "GET_EVENTS_SUCCESS",

  SELECT_EVENT = "SELECT_EVENT",
  RESET_DELETED_EVENT = "RESET_DELETED_EVENT",
  RESET_UPDATED_EVENT = "RESET_UPDATED_EVENT",

  REMOVE_EVENT_FROM_STATE = "REMOVE_EVENT_FROM_STATE",
}

export default Types;